ajout de langue

+++ MobileApp_Araignee/pages/home.js
@@ -1,31 +1,43 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, SafeAreaView, TouchableOpacity, FlatList } from 'react-native';
+import { View, Text, StyleSheet, SafeAreaView, ActivityIndicator, FlatList } from 'react-native';
 import Configuration from '../composants/Configuration';
-import { useSelector } from 'react-redux';
-import { collection, getDocs   } from 'firebase/firestore';
+import { collection, getDocs, deleteDoc, doc } from 'firebase/firestore';
 import { db } from '../api/firebase';
 import useBD from '../hooks/useBD'; 
+import StandardButton from '../components/StandardButton';
+import { Alert } from 'react-native';
+import { useTranslation } from 'react-i18next'; 
 
 const Home = () => {
   const [tempPrefs, setTempsPrefs] = useState([]);
-  const { prefs, getPrefsLocal, ajouterPrefsLocal, supprimerPrefsLocal, supprimerToutPrefsLocal } = useBD(); 
+  const [loading, setLoading] = useState(true);
+  //const { prefs, getPrefsLocal, ajouterPrefsLocal, supprimerPrefsLocal, supprimerToutPrefsLocal } = useBD(); 
+  const { t } = useTranslation(); 
 
   useEffect(() => {
     const getPrefs = async () => {
+      setLoading(true);
       try {
         const querySnapshot = await getDocs(collection(db, "comfort_preferences"));
         const list = [];
-        supprimerToutPrefsLocal();
+        //supprimerToutPrefsLocal();
         querySnapshot.forEach((doc) => {
           const data = { id: doc.id, ...doc.data() };
           list.push(data);
-          ajouterPrefsLocal(data.active, data.humidity, data.name, data.temperature);
+          //ajouterPrefsLocal(data.active, data.humidity, data.name, data.temperature);
+          console.log("test3");
         });
+        console.log("test4");
         setTempsPrefs(list);
+        console.log("test5");
       } catch (error) {
         console.log("Erreur lors de la récupération : " + error);
         getPrefsLocal();
         setTempsPrefs(prefs);
+        console.log("test6");
+      } finally {
+        setLoading(false); 
+        console.log("test7");
       }
     };
 
@@ -33,18 +45,32 @@ const Home = () => {
   }, []);
 
   const deletePrefs = async (id) => {
-    try {
-      await deleteDoc(doc(db, "comfort_preferences", id));
-
-      console.log("Document supprimé avec succès");
-      supprimerPrefsLocal(id);
-      getPrefs(); 
-    } catch (error) {
-      console.log("Erreur lors de la suppression : " + error);
-    }
+    Alert.alert(
+      t('home.confirm_delete_title'), 
+      t('home.confirm_delete_message'), 
+      [
+        {
+          text: t('no'), 
+          style: "cancel", 
+        },
+        {
+          text: t('yes'), 
+          onPress: async () => {
+            try {
+              await deleteDoc(doc(db, "comfort_preferences", id)); 
+              console.log("Document supprimé avec succès");
+              supprimerPrefsLocal(id); 
+              getPrefs(); 
+            } catch (error) {
+              console.log("Erreur lors de la suppression : " + error);
+            }
+          },
+          style: "destructive", 
+        },
+      ],
+      { cancelable: true } 
+    );
   };
-
-  const temperatureUnit = useSelector((state) => state.parameters.temperature_humidity_unit);
 
   const renderItem = ({ item }) => (
     <Configuration
@@ -52,9 +78,7 @@ const Home = () => {
       temperature={item.temperature}
       humidity={item.humidity}
       isActive={item.active}
-      onDelete={() =>{
-        deletePrefs(item.id);
-      }}
+      onDelete={() => deletePrefs(item.id)}
       onEdit={() => console.log("Editer : ", item.id)}
     />
   );
@@ -62,19 +86,25 @@ const Home = () => {
   return (
     <SafeAreaView style={styles.container}>
       <View style={styles.content}>
-        <Text style={styles.title}>Accueil</Text>
 
-        <TouchableOpacity style={styles.addButton}>
-          <Text style={styles.addButtonText}>Ajouter une configuration</Text>
-        </TouchableOpacity>
-
-        <View style={styles.configurationsContainer}>
-          <FlatList
-            data={tempPrefs}
-            renderItem={renderItem}
-            keyExtractor={(item) => item.id} 
-            ListEmptyComponent={<Text>Aucune configuration trouvée.</Text>}
+        <View style={{ marginTop: 30 }}>
+          <StandardButton
+              label={t('home.add_configuration')} 
+              color="green" 
           />
+        </View>
+        
+        <View style={styles.configurationsContainer}>
+          {loading ? (
+            <ActivityIndicator size="large" color="#4B4E6D" />
+          ) : (
+            <FlatList
+              data={tempPrefs === null || []}
+              renderItem={renderItem}
+              keyExtractor={(item) => item.id} 
+              ListEmptyComponent={<Text>{t('home.no_configurations_found')}</Text>} 
+            />
+          )}
         </View>
       </View>
     </SafeAreaView>
@@ -96,24 +126,8 @@ const styles = StyleSheet.create({
     marginTop: 20,
     color: '#4B4E6D',
   },
-  addButton: {
-    backgroundColor: '#84DCC6',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: 15,
-    marginTop: 40,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  addButtonText: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#4B4E6D',
-  },
   configurationsContainer: {
+    flex: 1,
     marginTop: 20,
   },
 });
